requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
bycrypt
python-dotenv
python-jose
python-multipart